submit_unit_tx returns the hash bytes (as bytes array)
and some refactoring
